+++ src/lib/common.ts
@@ -101,7 +101,6 @@ export const db = client.db("metahkg");
 
 export const threadCl = db.collection("thread");
 export const usersCl = db.collection("users");
-export const limitCl = db.collection("limit");
 export const verificationCl = db.collection("verification");
 export const categoryCl = db.collection("category");
 export const votesCl = db.collection("votes");