+++ src/routes/threads/thread/comments/index.ts
@@ -25,18 +25,18 @@ import emotions from "./emotions";
 import votes from "./votes";
 import deleteComment from "./actions/delete";
 import edit from "./actions/edit";
-import checkRemoved from "../../../../plugins/checkComment";
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
+import checkComment from "../../../../plugins/checkComment";
 
 export default function (
     fastify: FastifyInstance,
     _opts: FastifyPluginOptions,
     done: (e?: Error) => void
 ) {
-    fastify.addHook("preHandler", checkRemoved);
+    fastify.register(create);
+    fastify.addHook("preHandler", checkComment);
     fastify.register(comment);
     fastify.register(replies);
-    fastify.register(create);
     fastify.register(vote);
     fastify.register(images);
     fastify.register(emotions);