+++ src/routes/users/resend.ts
@@ -15,13 +15,13 @@
  along with this program.  If not, see <https://www.gnu.org/licenses/>.
  */
 
-import { RecaptchaSecret, verificationCl, limitCl } from "../../lib/common";
+import { RecaptchaSecret, verificationCl } from "../../lib/common";
 import { verifyCaptcha } from "../../lib/recaptcha";
 import { Static, Type } from "@sinclair/typebox";
-import Limit from "../../models/limit";
 import { FastifyInstance, FastifyPluginOptions, FastifyRequest } from "fastify";
 import { mg, mgDomain, verifyMsg } from "../../lib/mailgun";
 import { EmailSchema, RTokenSchema } from "../../lib/schemas";
+import { sha256 } from "../../lib/sha256";
 
 export default (
     fastify: FastifyInstance,
@@ -35,9 +35,25 @@ export default (
 
     fastify.post(
         "/resend",
-        { schema: { body: schema } },
+        {
+            schema: { body: schema },
+            config: {
+                rateLimit: {
+                    max: 2,
+                    ban: 5,
+                    // one day
+                    timeWindow: 1000 * 60 * 60 * 24,
+                    keyGenerator: (
+                        req: FastifyRequest<{ Body: Static<typeof schema> }>
+                    ) => {
+                        return sha256(req.body?.email);
+                    },
+                },
+            },
+        },
         async (req: FastifyRequest<{ Body: Static<typeof schema> }>, res) => {
             const { email, rtoken } = req.body;
+            const hashedEmail = sha256(email);
 
             if (!(await verifyCaptcha(RecaptchaSecret, rtoken)))
                 return res
@@ -45,31 +61,25 @@ export default (
                     .send({ statusCode: 429, error: "Recaptcha token invalid." });
 
             const verificationUserData = await verificationCl.findOne({
-                email,
+                type: "register",
+                email: hashedEmail,
             });
+
             if (!verificationUserData)
                 return res.code(404).send({ statusCode: 404, error: "Email not found." });
 
-            if (
-                (await limitCl.countDocuments({
-                    type: "resend",
-                    email,
-                })) >= 5
-            )
-                return res
-                    .status(429)
-                    .send({ error: "You can only resend 5 times a day." });
+            try {
+                await mg.messages.create(
+                    mgDomain,
+                    verifyMsg({ email, code: verificationUserData.code })
+                );
+            } catch {
+                return res.code(500).send({
+                    statusCode: 500,
+                    error: "An error occurred while sending the email.",
+                });
+            }
 
-            await mg.messages.create(
-                mgDomain,
-                verifyMsg({ email, code: verificationUserData.code })
-            );
-
-            await limitCl.insertOne({
-                type: "resend",
-                email,
-                createdAt: new Date(),
-            } as Limit);
             res.send({ success: true });
         }
     );